Empty the ruff report and make it a gate
466 findings that did not block. That is the worst of both worlds: a
permanently red tick is one nobody reads, so it protected nothing while looking
like it did.

389 OF THEM WERE TWO RULES THAT ARE WRONG HERE, and are now off in pyproject
with the argument at the switch. E501 fired 235 times, median 105 characters,
almost all of it prose in a comment or a message string -- which a formatter
could not split either, and ruff's own guidance is to leave line length to the
formatter. RUF001/2/3 fired 154 times and 137 were the dotless "i" in the
Turkish catalogue: not a suspicious lookalike, a different letter of the
language this interface speaks. A rule that flags correct Turkish 137 times
teaches people to stop reading the output.

THE OTHER 77 WERE FIXED. A zip that now says strict=False where it means it and
says why, two nested ifs that read better as one, an int(round(...)) rounding
something already an int, a closure that binds the dialog it uses instead of
reading it out of a loop.

FOUR WERE FIXED BY HAND BECAUSE THE AUTOMATIC FIX MADE THE CODE WORSE. RUF005
rewrites [a] + b as [*a, *b], and on a multi-line constructor call it does that
by inlining the whole call onto one 200-character line. It only looked
acceptable because E501 had just been switched off -- the same edit written by
hand is an improvement, applied by the tool it was vandalism with a green tick.

UP040 WOULD HAVE BROKEN THE APPLICATION SILENTLY, AND THE SUITE CAUGHT IT.
`type X = Literal[...]` is right for 32 of the 33 aliases and wrong for the four
read at RUN time: get_args returns an empty tuple for a PEP 695 alias.
BoardMaterial, BodyArchetype and ConductorKind stopped listing anything, so
three completeness tests began asserting that an empty set equals an empty set
-- checking nothing, in green. RoutingStyle was worse: MCP validates an agent's
requested style with `style not in get_args(RoutingStyle)`, which against an
empty tuple refuses every style there is, and neither call raises. All four keep
TypeAlias with the reason written above them.

The linter is pinned to a minor. An unbounded dependency took this project apart
twice this morning and a linter is the same hazard in a smaller way: rules
arrive with releases, and a gate that fails on a tree nobody touched is a gate
somebody switches off.

ruff format is still not run: 40 of 57 files rewritten and every line of blame
pointing at a reformat is a decision of its own, not a side effect of this one.

+++ src/perfstudio/ui/export_pdf.py
@@ -117,8 +117,8 @@ def verify_scale(scene: Any, board: Board, dpi: float = 300.0, span_holes: int =
     """
     px_per_mm = dpi / MM_PER_INCH
     outline = board_outline_mm(board)
-    img_w = int(round(outline.width * px_per_mm))
-    img_h = int(round(outline.height * px_per_mm))
+    img_w = round(outline.width * px_per_mm)
+    img_h = round(outline.height * px_per_mm)
 
     image = QImage(img_w, img_h, QImage.Format.Format_ARGB32)
     image.fill(QColor("#ffffff"))
